pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agnes"
version = "0.1.0"
description = "Agnes — Phase 1 metalayer foundation (DAG executor, three-DB store, audit/cost spine)."
requires-python = ">=3.12"
readme = "projectbriefing.md"
dependencies = [
    "aiosqlite>=0.19",
    "pyyaml>=6",
    "pydantic>=2.5",
    "anthropic>=0.40,<2",
    "httpx>=0.27",
    "python-dotenv>=1",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "rapidfuzz>=3",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
adk = ["google-adk>=1.29.0"]
pydantic_ai = ["pydantic-ai-slim[cerebras]"]
dev = ["pytest>=8", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]
